java-collections-ru/lambdas/src/test/java/exercise/AppTest.java
package exercise;
// BEGIN

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

import java.util.Arrays;

class AppTest {

    @Test
    void testEnlargeArrayImage1() {
        String[][] image = {
                {"+", "-"},
                {"-", "+"},
        };

        String[][] actual = App.enlargeArrayImage(image);
        String[][] expected = {{"+", "+", "-", "-"}, {"+", "+", "-", "-"}, {"-", "-", "+", "+"}, {"-", "-", "+", "+"}};
        assertThat(Arrays.deepToString(actual)).isEqualTo(Arrays.deepToString(expected));
    }

    @Test
    void testEnlargeArrayImage2() {
        String[][] image = {
                {"*", "*", "*", "*"},
                {"*", " ", " ", "*"},
                {"*", " ", " ", "*"},
                {"*", "*", "*", "*"},
        };

        String[][] actual = App.enlargeArrayImage(image);
        String[][] expected = {
                {"*", "*", "*", "*", "*", "*", "*", "*"},
                {"*", "*", "*", "*", "*", "*", "*", "*"},
                {"*", "*", " ", " ", " ", " ", "*", "*"},
                {"*", "*", " ", " ", " ", " ", "*", "*"},
                {"*", "*", " ", " ", " ", " ", "*", "*"},
                {"*", "*", " ", " ", " ", " ", "*", "*"},
                {"*", "*", "*", "*", "*", "*", "*", "*"},
                {"*", "*", "*", "*", "*", "*", "*", "*"},
        };

        assertThat(actual).isEqualTo(expected);
    }

    @Test
    void testEnlargeArrayImage3() {
        String[][] image = {
                {"*", "*", "*", "*"},
                {"*", " ", " ", "*"},
                {"*", " ", " ", "*"},
                {"*", "*", "*", "*"},
        };

        String[][] actual = App.enlargeArrayImage(image);
        int lengthActual = actual[0].length * actual.length;
        assertThat(lengthActual).isEqualTo(64);
    }

    @Test
    void testEnlargeArrayImage4() {
        String[][] image = {};
        String[][] actual = App.enlargeArrayImage(image);
        String[][] expected = {};
        assertThat(actual).isEqualTo(expected);
    }

    @Test
    void testEnlargeArrayImage5() {
        String[][] image = {
                {"@", "@"},
                {"@", "|"},
                {"—", "|"},
                {"@", "|"},
                {"@", "@"},
        };

        String[][] enlargedImage = {
                {"@", "@", "@", "@"},
                {"@", "@", "@", "@"},
                {"@", "@", "|", "|"},
                {"@", "@", "|", "|"},
                {"—", "—", "|", "|"},
                {"—", "—", "|", "|"},
                {"@", "@", "|", "|"},
                {"@", "@", "|", "|"},
                {"@", "@", "@", "@"},
                {"@", "@", "@", "@"},
        };

        String[][] actual = App.enlargeArrayImage(image);
        assertThat(actual).isEqualTo(enlargedImage);
    }

}
// END
